Interprete impported files in new child environments
New files are interpreted in full in a child environment, which means
they correctly have access to everything defined in the global
environment.

Because of this structure, definitions in the child environment don't
become available in the parent environment automatically. So once the
interpreter is done interpreting the file, we have to merge the child
environment into the parent. I am not sure this is the right solution
long term, but it gets the test suite passing at least! I can revisit
this later when other parts are more fleshed out and I have more
information.

Note: Because function definitions capture their defining environment
and use that to interprete the function body, no additional work is
needed to make sure that deep stack traces work, which I think is pretty
neat - I thought it would need a lot more work.